Updated Sleep time to include fractions of seconds

+++ src/Modules/Numbers.py
@@ -4,7 +4,7 @@ import time
 def numbers():
     try:
         count = int(input("Enter the number until which you want to spam : \n> "))
-        sleep = int(input("Enter time delay(in seconds) between each message : \n> "))
+        sleep = float(input("Enter time delay(in seconds) between each message : \n> "))
     except:
         print("ERROR : Enter Only Numbers")
         input("Press any key to exit ")
