+++ src/Modules/RageSpam.py
@@ -14,7 +14,7 @@ def message():
 def rage():
     try:
         count = int(input("Enter the number of spam messages you want to send : \n> "))
-        sleep = int(input("Enter time delay(in seconds) between each message : \n> "))
+        sleep = float(input("Enter time delay(in seconds) between each message : \n> "))
     except:
         print("ERROR : Enter Only Numbers")
         input("Press any key to exit ")
