+++ src/Modules/Static.py
@@ -5,7 +5,7 @@ def static():
     message = input("Enter the String you want to spam : \n> ")
     try:
         count = int(input("Enter the number of times you want to spam the message : \n> "))
-        sleep = int(input("Enter time delay(in seconds) between each message : \n> "))
+        sleep = float(input("Enter time delay(in seconds) between each message : \n> "))
     except:
         print("ERROR : Enter Only Numbers")
         input("Press any key to exit ")
